+++ data/index.js
@@ -1,5 +1,9 @@
 'use strict';
 const zonefile = require('./zonefile.2017.json');
+const hemispheres = require('./hemisphere');
+
+//assumed hemisphere, based on continent
+const southern = { Australia: true, Chile: true, Brazil: true, Antarctica: true };
 
 //compress timezone data by continent
 const unpack = obj => {
@@ -12,10 +16,15 @@ const unpack = obj => {
       all[tz] = obj[cont][city];
       if (typeof all[tz] === 'number') {
         all[tz] = {
-          o: all[tz],
+          o: all[tz]
         };
       }
       all[tz].tz = tz;
+      if (southern[cont] === true || hemispheres.south[tz]) {
+        all[tz].h = all[tz].h || 's';
+      }
+      //assume north, unless it says otherwise (sorry!)
+      all[tz].h = all[tz].h || 'n';
     });
   });
   //alias this one