+++ changelog.md
@@ -1,4 +1,8 @@
 # v1.0.0 :rocket:
 
-## v1.1.0
+## v1.2.0
 * adds isBetween() method
+
+## v1.3.0
+* adds `spacetime.whereIts()` method
+* actually implement 😓 season by hemisphere